pyproject.toml
[tool.poetry]
name = "weather"
version = "0.1.0"
description = ""
authors = ["Tyberius Prime <tyberius_prime@coonabibba.de>"]
license = "MIT"
readme = "README.md"
include = ["weather/back.png", "weather/font/Piscolabis-Regular.ttf"]

[tool.poetry.dependencies]
python = "^3.11"
pillow = "^10.2.0"
aiomqtt = "^1.2.1"
poetry-dynamic-versioning = "^1.2.0"
aiohttp = "^3.9.1"
rich = "^13.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
weather = "weather:main"


